Move 'InMemoryCacheFeedStore' to its own file in production
