+++ Tests/FeedStoreChallengeTests.swift
@@ -5,54 +5,6 @@
 import XCTest
 import FeedStoreChallenge
 
-class InMemoryCacheFeedStore: FeedStore {
-    final class Cache {
-        internal let feed: [LocalFeedImage]
-        internal let timestamp: Date
-        
-        init(feed: [LocalFeedImage], timestamp: Date) {
-            self.feed = feed
-            self.timestamp = timestamp
-        }
-    }
-    
-    private final class Key {
-        private let key: String
-        
-        init(_ key: String) {
-            self.key = key
-        }
-    }
-    
-    private var key: Key
-    private var memoryCache = NSCache<Key, Cache>()
-    
-    init(key: String) {
-        self.key = Key(key)
-    }
-    
-    func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-        if let _ = memoryCache.object(forKey: key) {
-            memoryCache.removeObject(forKey: key)
-        }
-
-        completion(nil)
-    }
-    
-    func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-        let cache = Cache(feed: feed, timestamp: timestamp)
-        memoryCache.setObject(cache, forKey: key)
-        completion(nil)
-    }
-    
-    func retrieve(completion: @escaping RetrievalCompletion) {
-        guard let cache = memoryCache.object(forKey: key) else {
-            return completion(.empty)
-        }
-        completion(.found(feed: cache.feed, timestamp: cache.timestamp))
-    }
-}
-
 class FeedStoreChallengeTests: XCTestCase, FeedStoreSpecs {
 	
 	func test_retrieve_deliversEmptyOnEmptyCache() {
